+++ src/main.js
@@ -5,8 +5,8 @@ import FilmCardElement from './view/film-card-elements-view';
 import ButtonShowMore from './view/button-show-more-view';
 import SiteStatTemplate from './view/site-statistick-card-view';
 import SortTemplate from './view/site-sort-view';
-import ExtraMostListView from './view/extra-list-view';
-import ExtraTopListView from './view/extra-list-view';
+import ExtraMostListView from './view/extra-most-list-view';
+import ExtraTopListView from './view/extra-top-list-view';
 import PopupTemplate from './view/site-popup-view';
 import NoFilmsView from './view/no-films-view';
 import { generateListFilm, getRandomTasks } from './mock/film-list';
